+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 
 ## [Unreleased]
 - [CM-1265] Improved Event Callbacks.
+- [CM-1227] Added support to close the conversation ViewController
 - Added Zendesk Integration
 - [CM-1167] Added a feature Custom Bot Name on Conversation screen through chat context 
 ## [6.7.4] - 2022-11-02Z